+++ src/01_type/change_zero_last_bits.rb
@@ -1,0 +1,2 @@
+def change_zero_last_bits(bits, num_of_bits)
+end
