--- src/01_type/remove_last_bits.rb
@@ -1,2 +1,0 @@
-def remove_last_bits(bits, num_of_bits)
-end
