+++ test/01_type/all_test.rb
@@ -4,4 +4,4 @@ require_relative 'string_format_test'
 require_relative 'round_down_diff_test'
 require_relative 'calc_tax_test'
 require_relative 'compare_length_test'
-require_relative 'remove_last_bits_test'
+require_relative 'change_zero_last_bits_test'
